TL74Functions/requirements.txt
azure-functions
azure-cosmos
azure-core
os
logging
sys
edgar
pandas
json